#9 Refactor user model and mapping logic; move DTO conversion to UserMapper and clean up User-related files
