+++ backend/src/main/kotlin/dev/ecckea/agilepath/backend/domain/user/model/User.kt
@@ -2,8 +2,6 @@ package dev.ecckea.agilepath.backend.domain.user.model
 
 import com.fasterxml.jackson.annotation.JsonTypeInfo
 import com.fasterxml.jackson.annotation.JsonTypeName
-import dev.ecckea.agilepath.backend.domain.user.dto.UserResponse
-import dev.ecckea.agilepath.backend.shared.utils.toZonedDateTime
 import java.time.Instant
 
 @JsonTypeInfo(use = JsonTypeInfo.Id.CLASS, include = JsonTypeInfo.As.PROPERTY, property = "@class")
@@ -17,14 +15,4 @@ data class User(
     val createdAt: Instant = Instant.now(),
     val modifiedAt: Instant? = null,
     val modifiedBy: String? // UserId
-)
-
-fun User.toDTO(): UserResponse = UserResponse(
-    id = id,
-    email = email,
-    fullName = fullName,
-    avatarUrl = avatarUrl,
-    githubUsername = githubUsername,
-    githubProfileUrl = githubUsername?.let { "https://github.com/$it" },
-    createdAt = toZonedDateTime(createdAt)
 )
