+++ backend/src/main/kotlin/dev/ecckea/agilepath/backend/domain/user/repository/entity/UserEntity.kt
@@ -1,14 +1,7 @@
 package dev.ecckea.agilepath.backend.domain.user.repository.entity
 
-import dev.ecckea.agilepath.backend.domain.user.model.User
-import dev.ecckea.agilepath.backend.shared.utils.nowInZone
 import jakarta.persistence.*
 import java.time.Instant
-import java.time.ZonedDateTime
-import java.util.*
-import com.fasterxml.jackson.databind.annotation.JsonSerialize
-import com.fasterxml.jackson.datatype.jsr310.ser.ZonedDateTimeSerializer
-import com.fasterxml.jackson.datatype.jsr310.ser.InstantSerializer
 
 @Entity
 @Table(name = "users")
@@ -40,15 +33,4 @@ class UserEntity(
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "modified_by")
     var modifiedBy: UserEntity? = null
-)
-
-fun UserEntity.toModel(): User = User(
-    id = id,
-    githubUsername = githubUsername,
-    email = email,
-    fullName = fullName,
-    avatarUrl = avatarUrl,
-    createdAt = createdAt,
-    modifiedAt = modifiedAt,
-    modifiedBy = modifiedBy?.id,
 )
